New method: ActionView::Helpers::CacheHelper#conditionally_cache.
condtionally_cache takes a new first argument that is evaluated in a boolean
context. When true, it passes control off to the usual #cache method to for
fragment caching. When false, the block is captured and rendered normally,
but without fragment caching.

This is useful for things like paginated results, where you might want to
fragment cache the first page, but not pages 2+.