New argument to fragment macros to eval key in instance context.
If the fragment key is made up of something that can only be evaluated in
the context of the test instance (e.g. it includes the ID of a record
created by the context setup), the key can't be a simple string. The
fragment macros now accept either a key or a block argument, and
use the value of the block if one is provided. (Technically, you could
supply both a key and a block, but the key is ignored in this case.)